* all .h, .m           float -> CGFloat, int -> NSInteger transitions
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2225 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Games/Jigsaw/Document.h
@@ -22,12 +22,12 @@ typedef enum {
     NSMutableArray *clusters;
     NSView *view;
     NSString *nameOfImageFile;
-    int px, py;
+    NSInteger px, py;
 
     BOOL solving;
-    int done;
+    NSInteger done;
 
-    int piece_width, piece_height;
+    NSInteger piece_width, piece_height;
 }
 
 + actionMenu;
@@ -35,7 +35,7 @@ typedef enum {
 - init;
 - (void)dealloc;
 
-- (int)setDone:(int)flag;
+- (NSInteger)setDone:(NSInteger)flag;
 
 - (NSSize)withPadding;
